+++ cmd/toglacier/toglacier.go
@@ -224,7 +224,7 @@ func commandSync(c *cli.Context) error {
 		logger.Out = ioutil.Discard
 	}
 
-	if err := toGlacier.Backup(config.Current().Paths, config.Current().BackupSecret.Value); err != nil {
+	if err := toGlacier.Backup(config.Current().Paths, config.Current().BackupSecret.Value, float64(config.Current().ModifyTolerance)); err != nil {
 		logger.Error(err)
 	}
 
@@ -305,7 +305,7 @@ func commandList(c *cli.Context) error {
 func commandStart(c *cli.Context) error {
 	scheduler := gocron.NewScheduler()
 	scheduler.Every(1).Day().At("00:00").Do(func() {
-		if err := toGlacier.Backup(config.Current().Paths, config.Current().BackupSecret.Value); err != nil {
+		if err := toGlacier.Backup(config.Current().Paths, config.Current().BackupSecret.Value, float64(config.Current().ModifyTolerance)); err != nil {
 			logger.Error(err)
 		}
 	})