+++ doc.go
@@ -1,0 +1,2 @@
+// Package toglacier have all the functions to manage your backups in the cloud.
+package toglacier